split methods that took user inputs (inluding prompts) away from UserInterface.cs and into UserInput.cs
