+++ Console.CodingTracker.oksanatk/CodingSession.cs
@@ -1,4 +1,4 @@
-﻿namespace TSCA.CodingTracker;
+﻿namespace CodingTracker.oksanatk;
 
 internal class CodingSession
 {
